[SOL] CI: run tests on push and improve run times
Run tests on push to master in addition to running them on pull requests and
setup S3 caching to improve run times.

Docker images for for the jobs are cached in s3://cached-ci-artifacts/docker
and LLVM/lld build artifacts are cached through sccache.

A custom sccache fork is used. The custom fork is needed to workaround
https://github.com/mozilla/sccache/issues/633 happening with our S3 bucket,
which only supports the v4 signature scheme.